moved SceneLoaded from SceneManager to Scene, instancing renamed stuff for clarity, trnasform handle picking

+++ tofuProject/Assets/Shaders/ModelMousePicking.glsl
@@ -15,14 +15,14 @@ layout (location = 9) in float a_id;
 layout (location = 10) in vec2 a_uv_offset;
 uniform mat4 u_viewProjection;
 
-flat out float v_id;
+flat out uint v_id;
 
 void main(void)
 {
 	mat4 a_model = mat4(vec4(a_model_1, 0), vec4(a_model_2, 0), vec4(a_model_3, 0), vec4(a_model_4, 1));
 	mat4 mvp = u_viewProjection * a_model;
 
-	v_id = a_id;
+	v_id = uint(a_id);
 
 	gl_Position = mvp * vec4(a_pos.xyz, 1.0);
 }
@@ -31,16 +31,20 @@ void main(void)
 #version 410 core
 //layout (location = 0) out uint fragColor;
 layout (location = 0) out vec4 fragColor;
-flat in float v_id;
+flat in uint v_id;
 void main(void)
 {
-//	fragColor = uint(v_id);  must be integer framebuffer
-//	fragColor = vec4(v_id, 0.0, 0.0, 1.0);
-//float r = float((int(v_id) & 0xFF000000) >> 24) / 255.0;
-//float g = float((int(v_id) & 0x00FF0000) >> 16) / 255.0;
-//float b = float((int(v_id) & 0x0000FF00) >> 8) / 255.0;
-//float a = float(int(v_id) & 0x000000FF) / 255.0;
+//	float a = float((v_id >> 24) & 0xFFu) / 255.0; // Extract alpha (highest byte)
+//	float r = float((v_id >> 16) & 0xFFu) / 255.0; // Extract red (highest byte)
+//	float g = float((v_id >> 8) & 0xFFu) / 255.0;  // Extract green (middle byte)
+//	float b = float(v_id & 0xFFu) / 255.0;         // Extract blue (lowest byte)
 //
-//fragColor = vec4(r, g, b, a);
+//	fragColor = vec4(r, g, b, a); // RGB color with alpha = 1.0
+
 fragColor = vec4(int(v_id)/255.0,0,0,0);
+
 }
+
+
+
+
